pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "iris-vector-graph"
version = "1.1.6"
description = "Knowledge graph and vector search platform on InterSystems IRIS for financial services and biomedical applications"
readme = "README.md"
license = "MIT"
authors = [
    {name = "Thomas Dyar", email = "thomas.dyar@intersystems.com"}
]
maintainers = [
    {name = "Thomas Dyar", email = "thomas.dyar@intersystems.com"}
]
keywords = [
    "intersystems-iris",
    "vector-search",
    "knowledge-graph",
    "rag",
    "biomedical",
    "fraud-detection",
    "graph-database",
    "protein-interactions",
    "bitemporal",
    "hnsw"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Financial and Insurance Industry",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Database",
    "Topic :: Database :: Database Engines/Servers",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
requires-python = ">=3.11"

# Core dependencies - everything needed for graph operations, APIs, and demos
dependencies = [
    "intersystems-irispython>=3.2.0",
    "networkx>=3.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.118.0",
    "uvicorn>=0.37.0",
    "pydantic>=2.11.9",
    "httpx[http2]>=0.28.1",
    "strawberry-graphql[fastapi]>=0.280.0",
    # Demo server dependencies
    "python-fasthtml>=0.8.1",
    "faker>=28.0.0",
]

# Optional dependencies for different use cases
[project.optional-dependencies]
# Machine learning features (fraud detection models, neural networks)
ml = [
    "torch>=2.0.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0"
]

# Development tools
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "playwright>=1.40.0"
]

# Performance profiling
performance = [
    "psutil>=5.9.0",
    "memory-profiler>=0.61.0"
]

# Visualization tools
visualization = [
    "matplotlib>=3.7.0",
    "plotly>=5.15.0",
    "graphviz>=0.20.0"
]

# Biomedical data sources
biodata = [
    "biopython>=1.81",
    "bioservices>=1.11.0",
    "mygene>=3.2.0"
]

[project.urls]
Homepage = "https://github.com/intersystems-community/iris-vector-graph"
Documentation = "https://github.com/intersystems-community/iris-vector-graph/tree/main/docs"
Repository = "https://github.com/intersystems-community/iris-vector-graph"
Issues = "https://github.com/intersystems-community/iris-vector-graph/issues"

[project.scripts]
networkx-loader = "scripts.ingest.networkx_loader:main"

[tool.hatch.build.targets.wheel]
packages = ["iris_vector_graph"]

# Tool configurations
[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "iris",
    "networkx",
    "plotly.*",
    "bioservices.*",
    "mygene",
    "fasthtml.*",
    "faker"
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--disable-warnings"
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "performance: marks tests as performance benchmarks",
    "requires_iris: marks tests that require IRIS database connection",
    "requires_database: marks tests that require database connection (alias for requires_iris)",
    "e2e: marks tests as end-to-end system tests (all query engines)"
]

[tool.coverage.run]
source = ["scripts", "tests"]
omit = [
    "*/tests/*",
    "*/venv/*",
    "*/.venv/*",
    "*/build/*",
    "*/dist/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:"
]
